+++ Assets/NatSuite/NatShare/Plugins/Managed/ISharePayload.cs
@@ -16,19 +16,19 @@ namespace NatSuite.Sharing {
         /// <summary>
         /// Add text to the payload.
         /// </summary>
-        ISharePayload AddText (string text);
+        void AddText (string text);
 
         /// <summary>
         /// Add an image to the payload.
         /// </summary>
         /// <param name="image">Image to the shared.</param>
-        ISharePayload AddImage (Texture2D image);
+        void AddImage (Texture2D image);
 
         /// <summary>
         /// Add a media file to the payload.
         /// </summary>
         /// <param name="path">Path to media file to be shared.</param>
-        ISharePayload AddMedia (string path);
+        void AddMedia (string path);
 
         /// <summary>
         /// Commit the payload and return the success value.